+++ srcs/game/utils_2.c
@@ -6,18 +6,18 @@
 /*   By: asayad <asayad@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/24 21:43:32 by hmoukit           #+#    #+#             */
-/*   Updated: 2025/01/14 11:27:56 by asayad           ###   ########.fr       */
+/*   Updated: 2025/01/18 21:25:12 by asayad           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <cub3d.h>
 
-float	deg2rad(float angle_deg)
+double	deg2rad(double angle_deg)
 {
 	return (angle_deg * PI / 180);
 }
 
-float	rad2deg(float angle_rad)
+double	rad2deg(double angle_rad)
 {
 	return ((angle_rad / PI) * 180);
 }
@@ -27,7 +27,7 @@ int get_rgba(int r, int g, int b, int a)
     return (r << 24 | g << 16 | b << 8 | a);
 }
 
-float	square(float i)
+double	square(double i)
 {
 	return (i * i);
 }
